requirements.txt

beautifulsoup4 
bs4
frozendict
geopy
ipython
kafka-python
matplotlib
numpy
py4j
pyelasticsearch
pymongo
requests
scipy
selenium
scikit-learn
tabulate
tldextract
wikipedia
findspark
iso8601
urllib3==1.24
joblib
flask==2.2.5
flask-socketio==5.3.6
python-socketio==5.7.2
python-engineio==4.3.2
eventlet==0.33.3
cassandra-driver



